[DEV] Correctly use timezone concept
Currently when a controller set the datetime with the SDK (currentDate &
currentTime cmds), using the ISO 8601 datetime standard, the timezone is
completly ignored and the system time is set as if the given time is utc+0.

For instance if the controller gives T14:00:00+0200
(=14h CEST) (=14h UTC+2) (=12h UTC+0)

Currently the drone set the system time to 14h00 UTC+0.
This is an error.

The drone should set the system time to 12h00 UTC+0 and save the timezone
(as an utc offset).

This patch correctly set the time to UTC+0 and save the timezone in
a boxinit property "timezone.utc_offset.sec" as an offset from UTC.

For instance:
Central Standard Time:         CST | UTC-6 | -21600
Central European Summer Time: CEST | UTC+2 |   7200
Coordinated Universal Time:    UTC | UTC+0 |      0