[DEV] move the use of the utc_offset
The utc_offset is now properly apply in the 'time_local_to_tm' function
instead of the 'time_local_format' function.

Because it is not the role of the 'time_local_format' function to do this.
And the 'time_local_to_tm' was wrong if called externally without doing
the offset modification.