[DEV] random: amortize /dev/urandom read cost
futils_random_uint{8,16,32,64} return values from a
CSPRNG, aka. cryptographically secure random number
generator, that is, a well seeded, secure RNG that
don't leak its internal state in its outputs, and
prevent recovering previous generated values, even
if its internal state is leaked somehow.

To do that, on Unix, futils_random_uint{8,16,32,64}
open, read, and close "/dev/urandom" for each random
number generated.

When not used for security purpose, where unpredictability
is not much of a concern, going through "/dev/urandom"
is really costly and quite overkill especially for
8bits, 16bits values, that cannot 'contain' much
entropy.

A faster CSPRNG would be useful to generate small
values.

As the operating system CSPRNG is fast enough when
generating large chunk of random data, this commit
implements a pool where random data are made available
to pick as needed to generate small values, reducing
the amount of open(), read(), close() syscalls made.

To be usuable in multithread environment, the pool
is shared between all threads, protected by a pthread
mutex.

Below is a crude comparison between previous implementation
and the new one, made on a x86_64 computer running
Ubuntu 18.04, made with dedicated test programs.
Results are in MiB/s:

                 futils_random8()  futils_random_bytes()
urandom               0.347               220.020
getrandom             1.378               225.376
getrandom, pool      33.649               219.940

It should be noted the cost of lock/unlocking is be
quite high regarding the size of the value to extract,
thus generating small random values is not so optimized
compared to generating large buffer.

Pool buffer size was fixed to 512 bytes as it seems
to be the sweet spot on the computer used for the
benchmarks:

                                           PRNG speed
    +-------------------------------------------------------------------------------------+
    |     ++  +   +         +                  +                                     +    |
    |                                                                   getrandom ***A*** |
    |                                                                                     |
    |                                                             *******************A    |
    |                                 *********A******************                        |
 35 |-+                   **A*********                                                  +-|
    |                *****                                                                |
    |             A**                                                                     |
    |           **                                                                        |
    |          *                                                                          |
    |         A                                                                           |
    |         *                                                                           |
    |        *                                                                            |
    |        *                                                                            |
    |       *                                                                             |
 30 |-+     *                                                                           +-|
    |      A                                                                              |
    |      *                                                                              |
    |      *                                                                              |
    |      *                                                                              |
    |      *                                                                              |
    |      *                                                                              |
    |     *                                                                               |
    |     *                                                                               |
    |     *                                                                               |
    |     *                                                                               |
 25 |-+   A                                                                             +-|
    |                                                                                     |
    |     ++  +   +         +                  +                                     +    |
    +-------------------------------------------------------------------------------------+
         64  256 512      1024               2048                                  4096
          128                            pool size (bytes)

BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE

The pool *is not* flushed on fork(), so child and
parent will produces the same random numbers until
they have consumed their respective random data pool.

The values from the pool are erased as they're
consumed, so a leak of the buffer doesn't reveal
previous values. But it would reveal all future values
returned by futils_random{8,16,32,64}() and futils_random_bytes()
(for length compatible with the pool) before the pool
is reloaded.

In a multithreaded process, it's not safe to call any
futils_random...() functions in the child process after
calling fork(): in the parent process a thread might have
locked the pool while another thread called fork(), then
in the child, the lock cannot be released, thus calling
futils_random_...() here will deadlock.

Also, even in a single threaded process, it's not safe
to call any futils_random...() functions in a signal
handler, as 1) pthread_mutex_lock() and
pthread_mutex_unlock() are not "async signal safe", and
2) the pool can be locked by the process before it is
interrupted by the signal, thus calling futils_random_...()
in the signal handler will deadlock.

BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE BEWARE