[DEV] random: add tool to exercise and benchmark PRNG
Add a program to test the output and speed of the PRNG:
it outputs the generate binary stream on standard
output, thus it can be fed to tools like dieharder for
evaluation, or pv for benchmark, and, on exit, after
fixed amount of bytes generated or when interrupted,
it will report generation speed in GiBytes, MiBytes,
KiBytes per seconds on standard error.

Usage:

1) produces an unlimited binary stream:

  ./futils-random random8 | hexdump -C

2) produces a bounded binary stream:

  ./futils-random random8 1048576 > 1M

3) using timeout tool, it can be stopped after
   some amount of time:

   timeout 60s ./futils-random random8 > /dev/null

The following "generators" are available

 - random8:      - use futils_random8()
 - random16:     - use futils_random16()
 - random32:     - use futils_random32()
 - random64:     - use futils_random64()
 - random-bytes: - use futils_random_bytes()
 - memset:       - use memset(), thus return only 0s
 - noop:         - no operation, returns only 0s too

The last two "generators" can be used to establish
benchmark baseline. They're mostly of interest to
have an idea of the fastest computation achievable
on a given target.