[DEV] random: retry reading /dev/urandom when interrupted
It's possible for read() to be interrupted by a signal,
in such case, read() can return -1 and set errno to EINTR,
or read() can return a smaller size than requested.

Instead of returning an error, it would be better to retry
the read until the buffer is filled or a fatal error happen.