[DEV] random: add simpler functions to retrieve random number
The API exposed by futils's random requires to check
the return value of each function. For example:

    uint64_t u64;

    if (futils_random64(&u64) != 0)
        /* error branch */

But since, the internal PRNG was modified to deal with
seeding error as critical, fatal, terminal error, it
always return success (except for parameter errors).
So, in almost all cases, testing for error is no more
required.

It would make it easier to have function that don't
return but the return value instead, such as the
following:

    uint64_t u64 = futils_randomr64();

This patch adds such new functions, with a "r" in the
name to not conflict with the current ones.

The existing functions still have merits. In particular
they ensure the types match, so it's not possible to
generate a 8bits random value stored in a 64bits variable
by mistake.

Also existing codes rely on them having the int (*)(uint{8,16,32,64}_t *)
prototypes, so changing them would break those.