[DEV] random: allow reseeding the PRNG (to support fork())
As noted in commit implementing the PRNG, a child
process will generate the same number sequence than
its parent.

To cope with such situation, this patch adds a
function that can be called in child process to
trigger reseeding the PRNG.

This function can be used safely in single threaded
program to reseed the PRNG in the child process.

Unfortunately, in multithreaded program, when pthread
thread specific data based implementation is used,
it's not safe to call any futils_random_*() functions
in the child process because of uses of so called
async signal unsafe functions.

In the other hand, when Thread Local Storage (TLS)
implementation is used, futils_random_*() functions
can be called safely in all cases.

As each thread have a dedicated pool, not accessible
to other threads, it's not possible for futils_random_reseed()
to actually reseed those states. So a global counter
is maintained by the PRNG, which is incremented each
time futils_random_reseed() is called. Then, next call
to futils_random_bytes() or futils_random{8,16,32,64}()
will test the counter and, if it's different from the
one recorded in current state, the function will
proceed with the reseeding, recording the current
counter value. The counter, aka. era is kept 'odd'
as all time, to be sure current era is never 0,
allowing to distinguish unseeded pool, even in the
unlikely case era overflow.
Atomics operations are used to not have to maintain
a lock to protect the global counter, which would
defeat the whole advantage of TLS. If C11's <stdatomic.h>
is not available, GCC atomics builtins are used (used
on Clang too). It should be noted GCC and Clang expose
<stdatomic.h> even with older C standards, so it's
probed at preprocessing time using GCC/Clang
__has_include() extension.

Usage:

    pid_t pid = fork();
    switch (pid) {
    case -1:
        perror("fork()";
        exit(EXIT_FAILURE);
    case 0:
        futils_random_reseed();
        child();
        break;
    default:
        parent(pid);
        break;
    }

Note: if the child process is created only to call exec(),
there's no need for the PRNG to be reseeded.