[DEV] random: add functions to return value up to maximum
Add functions to be used to return a value in a given
range, preventing usage of modulo ('%' operator) which
would introduce a slight bias on the generated value,
making the random variable more uniformly distributed.

For example, the following introduces a bias toward the
lowest values:

  futils_random16(&v);
  v %= 6;

See

https://ericlippert.com/2013/12/16/how-much-bias-is-introduced-by-the-remainder-technique/

or

https://stackoverflow.com/questions/10984974/why-do-people-say-there-is-modulo-bias-when-using-a-random-number-generator

To help preventing bias when using PRNG output, this
commit adds functions that use rejection method to
ensure the generated values have uniform distribution
(same probabilities): if random value is bigger than
the maximum, draw another one.

As it could require many attempts for small maximum,
and because reloading the pool could be an heavyweight
operation:

1) implementation extracts from the random pool only
   the minimum required bytes to hold the maximum value
   requested. Anyway as it's byte oriented, at worst
   7bits are discarded each time a random number is
   extracted from the pool.

2) the obtained random value is folded to the smallest
   range multiple of the PRNG, then compared to the
   maximum.

See also https://www.pcg-random.org/posts/bounded-rands.html

In the worst case, when maximum is a power of two,
each draw has a 50% chance of being rejected.

In the best case, when the maximum is a power of two
minus 1, each draw has almost 0% chance of being
rejected.

An unit test is added to check the added functions
actually return numbers in the [0..maximum] range.
This unit test doesn't check distribution of the
returned values, so you will have to trust the
de-biasing method used here. Anyway those tests
are (and were enough) to catch implementation
mistakes.