[DEV] random: add a function to emit a base64 string
Add a function to generate a string of random binary
data already encoded in base64.

This function would be used to build identifiers with
improved entropy density over hexadecimal encoded one.

Usage:

  char base64[((8 + 2) / 3) * 4 + 1];

  futils_random_base64(base64, sizeof(base64), 8);

  printf("%s", base64);

The function is given a buffer, buffer length, and
the number of random bytes to be encoded in base64.
The actual number of character stored in the buffer
would be ((count + 2) / 3) * 4.

Except if it's 0 byte long, buffer is always NUL
terminated, even if it's too short to hold the actual
number of character given above. In the later case,
the output is truncated.

The function returns the number of characters that
would be printed if buffer was big enough, like
snprintf(). Like snprintf() it returns an 'int',
because Windows doesn't have a ssize_t type like
in POSIX.