[DEV] random: make initial seeding error critical and fatal
With the ChaCha20 based CSPRNG, there's only one
reason for a failure: an error during initial
seeding.

If the operating system cannot provide the
requested seed, the system is in a broken state,
and no random number can be generated.

Not being able to seed the internal CSPRNG from
the operating system CSPRNG is a critical error
that cannot be recovered.

So this commit simplify implementation by handling
the only possible error as fatal for the program.

In practice this will never happen.