[DEV] random: add a function to random shuffle sort an array
Add a function to randomly shuffle the content of
an array using Fisher–Yates shuffle method.

The function use a qsort()-like prototype and can
shuffle any reasonable size of items / array.

As an optimization, it uses a small temporary
buffer. If the item's array are too big, a bigger
buffer is allocated with malloc().

See https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle
    https://lemire.me/blog/2016/06/30/fast-random-shuffling/

The unit test ensure every items in an array has,
at least once, be seen in each position. As it's
random, there's a chance this never happen. As
the number of possible permutations is n!, trying
to generate every one is impracticable. Hopefully,
it's an application of the coupon collector's
problem, so with only a (relatively) small number
of permutations generated, there's a great confidence
for the test to succeed.
In practice, the number of attempts is set to be
large enough that the test always succeed, in an
acceptable time.