[DEV] random: add functions to return double and float
Add functions to return random values as double or float
in the unit interval [0..1).

It uses the most simple method providing some assurance
regarding the uniform distribution of the floating point
values generated, as described here:
- http://vigna.di.unimi.it/xorshift/#remarks

The function will generate random values multiple of
- 1/2**53 for double
- 1/2**24 for float

Please note more complex solutions exist that can be used
to generate any double/float in range [0..1):
- https://mumble.net/~campbell/2014/04/28/uniform-random-float
- https://experilous.com/1/blog/post/perfect-fast-random-floating-point-numbers

The unit test only check that the double and float values
are in the expected range and every bits of their mantissa
are toggled. No check for actual randomness or uniform
distribution is implemented here.