[DEV] random: add a function to emit a random string
Add futils_random_string() function to generate a
random string from a given alphabet.

For example, to produce a 16 characters long string
from "[A-Z][a-z][0-9]" alphabet:

    futils_random_string(buffer, sizeof(buffer),
                         16,
                         "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
                         "abcdefghijklmnopqrstuvwxyz"
                         "0123456789");

futils_random_string() takes care of avoiding the
modulo bias that would have made the first characters
in the alphabet to appears slightly more often than
the others.

The function parameters and return value are following
the pattern established with futils_random_base16() and
futils_random_base64(), where
- the buffer length and the requested count are
  distinct parameters;
- the return value is the requested count as if buffer
  length was large enough, as does snprintf().

The unit test mostly checks for buffer overflow, but
also attempts to verify every characters in the alphabet
is used.

On the other hand, the test doesn't check distribution
of the characters in generated strings. But, by not
using modulo (%), futils_random_string() is immune to
the modulo bias issue.