[DEBUG] tests/dynmbox: Do not rely on implementation details
Unit tests testing the API rely on the fact that the current underlying
implementation is based on an UNIX pipe. Ideally, those tests should
limit themselves to testing that the API exhibit the expected behavior
described in the documentation, independently of how such API was
implemented.
This commit re-implements an internal function to rely on the existing
public API instead, as well as remove some tests which are not useful in
validating the API behavior.
Such tests will break anyway with the upcoming dynmbox re-implementation.