[BUGFIX] dynmbox re-implementation
Re-implementation of the dynmbox for Windows and Linux using pthread
locking primitives to ensure read/write atomicity, which the
implementation using pipes could not provide.

This fixes a bug when sending messages grater than PIPE_BUF bytes in
a multi-threaded program, which would cause dynmbox_peek() to report
truncated or corrupted messages, and potentially corrupting memory by
writing beyond the destintion buffer.
This also adds a Windows implementation of the dynmbox.

This new implementation uses a circular buffer to implement the mail box
queue, and standard pthread locking promitives to ensure operation
atomicity, regardless of the message length.