[DEV] timetools: add time_get_realtime()
Add a function to retrieve the current UTC time
expressed as a struct timespec. It's meant to
wrap clock_gettime(CLOCK_REALTIME), and emulate
it when not available:

On macOS and iOS, gettimeofday() is used if clock_gettime()
and CLOCK_REALTIME are not available (before macOS 10.12).

On Windows, FILETIME is used and converted to an Unix
timespec.

On ThreadX, the function is not implemented as it seem
there's so such things as a real-time clock available.