[DEV] random: make futils_random_strong() stronger
Under Linux, futils_random_strong() makes use of getrandom()
with O_NONBLOCK, or /dev/urandom if getrandom() is not
available or available but failed to fill the requested
buffer.

Without GRND_NONBLOCK, getrandom() would block until
the kernel CSPRNG is securely initialized from entropy
sources, but when used with GNRD_NONBLOCK, it just
returns -1 and set errno to EAGAIN, then futils_random_strong()
defaults to using /dev/urandom.

But using /dev/urandom is not so strong because it can
returns bytes before the kernel CSPRNG is fully initialized,
so those bytes may not be "cryptographically secure".
This is reported in following kernel logs:

    random: <comm>: uninitialized urandom read (XX bytes read)

For futils_random_strong() to stand for its name,
getrandom() must not be used with GRND_NONBLOCK, so that
it waits for the kernel CSPRNG to be initialized.

For systems without getrandom(), /dev/random must be used
instead of /dev/urandom for the same reason: to wait for
the kernel CSPRNG to be initialized. Then, to be consistent,
getrandom() should also be used with GRND_RANDOM, so that
it behaves like /dev/random.

Unfortunately, before Linux v5.6, getrandom(,, GRND_RANDOM)
and /dev/random also waits for its pool to contains enough
"entropy" to fulfill requests and this can lead to long
wait if the pool takes a long time to replenish. But this
is the price to pay for being strong.

Hopefully, since Linux v5.6, there's no more blocking
entropy pools, so getrandom(,, GRND_RANDOM) behaves like
getrandom(,, 0), turning GRND_RANDOM into a no-op, and
/dev/random behaves like getrandom(,,0), eg. blocking only
until kernel CSPRNG is securely initialized. So with newer
kernel, having futils_random_strong() using getrandom(,,GRND_RANDOM)
and /dev/random is a easy choice.