[DEV] random: add a function to emit a base32 string
Add a function to generate a string of random binary
data already encoded in Base32, as defined by RFC 4648.

This function would be used to build identifiers with
improved entropy density over hexadecimal encoded one.

Usage:

    /*
     * to prevent truncation, buffer should be at least
     * ((X + 4) / 5) * 8 + 1, where X is the number of
     * random bytes to encode in base32
     */
    char base32[64];

    futils_random_base32(base32, sizeof(base32), 32);

    printf("%s", base32);

The function is given a buffer, buffer length, and
the number of random bytes to be encoded in base32.
The actual number of character stored in the buffer
would be ((count + 4) / 5) * 8.

Except if it's 0 byte long, buffer is always NUL
terminated, even if it's too short to hold the actual
number of character given above. In the later case,
the output is truncated.

The function returns the number of characters that
would be printed if buffer was big enough, like
snprintf(). Like snprintf(), it returns an 'int',
because Windows doesn't have a ssize_t type like
in POSIX.