[DEV] random: emit valid padded base64 string
futils_random_base64() emits incorrect base64 when
padding.

To encode a single byte, it is first split in 6 bits
groups, and the last partial group is padded on the
right with 0:

    XXXXXX XX0000

To encode two bytes, they are first split in 6 bits
groups, and the last partial group is padded on the
right with 0:

    XXXXXX XXYYYY YYYY00

futils_random_base64() doesn't pad with 0, but with
random bits, that are, at best, ignored by Base64
decoder, but, at worst, made the whole string rejected
by stricter Base64 decoder.

For example, it was possible for the function to
emit a string such as

 //==

 111111 111111

This would be decoded as 0xff value, ignoring the last
4 bits. But 0xff would have been encoded correctly a
the string

 /w==

For a two characters padding string, the only valid
last characters are:

  .A==
  .Q==
  .g==
  .w==

For a single character padding string, the only valid
end are:

  ..A=
  ..E=
  ..I=
  ..M=
  ..Q=
  ..U=
  ..Y=
  ..c=
  ..g=
  ..k=
  ..o=
  ..s=
  ..w=
  ..0=
  ..4=
  ..8=

Addressing this issue is a matter of masking the low order
bits of the last random value to form the partial 6bits
group.